parallaxDrawer: merges trip creator back into one file
