+++ compnents/reusables/parallaxDrawer/index.tsx
@@ -16,16 +16,15 @@ import Animated, {
 import {
   GestureHandlerRootView,
 } from 'react-native-gesture-handler';
-import Svg, { Path } from 'react-native-svg';
+import { Path } from 'react-native-svg';
 import { moderateScale } from 'react-native-size-matters';
 import * as S from './styles';
 import ButtonIcon from '../buttonIcon';
 import { Gesture, GestureDetector } from 'react-native-gesture-handler';
 
 
-const { width: SCREEN_WIDTH, height: SCREEN_HEIGHT } = Dimensions.get('window');
+const { width: SCREEN_WIDTH, height: SCREEN_HEIGHT } = Dimensions.get('screen');
 const HALF_HEIGHT = SCREEN_HEIGHT / 2;
-const svgYOffset = SCREEN_HEIGHT > 800 ? -10 : 0;
 
 type ParallaxDrawerProps = {
   headerImage: any;
@@ -39,6 +38,10 @@ const ParallaxDrawer = ({ headerImage, children, onClose }: ParallaxDrawerProps)
   const TOP_SECTION_HEIGHT = moderateScale(70); // Define top section height
   const startY = useSharedValue(0);
 
+  if(translateY.value === 0){
+    translateY.value = HALF_HEIGHT
+  }
+ 
   const panGesture = Gesture.Pan()
     .onStart((event) => {
       startY.value = translateY.value;
@@ -98,68 +101,73 @@ const ParallaxDrawer = ({ headerImage, children, onClose }: ParallaxDrawerProps)
     };
   });
 
-  console.log(children)
+
+  const closeParallax = () => {
+    translateY.value = withTiming(0, { duration: 200 }, (isFinished) => {
+      if (isFinished) {
+        startY.value = 0;
+        translateY.value = 0;
+      }
+    });
+      onClose();
+  }
+  
 
   return (
-    <GestureHandlerRootView style={styles.container}>
+    <GestureHandlerRootView>
 
     <S.SafeArea>
       <S.BackButtonWrapper>
         <ButtonIcon
           icon="chevron-left"
-          onPress={onClose}
+          onPress={closeParallax}
           size="small"
         />
       </S.BackButtonWrapper>
       </S.SafeArea>
-      <Animated.View style={[styles.backgroundContainer, animatedBackgroundStyle]}>
-        <ImageBackground
-          source={headerImage}
-          style={styles.backgroundImage}
-          resizeMode="cover"
-        />
-      </Animated.View>
+      <S.BackgroundContainer>
+        <Animated.View style={[StyleSheet.absoluteFill, animatedBackgroundStyle]}>
+          <ImageBackground
+            source={headerImage}
+            style={StyleSheet.absoluteFill}
+            resizeMode="cover"
+          />
+        </Animated.View>
+      </S.BackgroundContainer>
 
       <GestureDetector gesture={panGesture}>
         <Animated.View style={[styles.drawer, animatedDrawerStyle]}>
-          <View style={styles.topTransparentSection}>
-            <Svg
+          <S.TopTransparentSection>
+            <S.StyledSvg
               width="100%"
               height={SCREEN_WIDTH * (320 / 1440)}
               viewBox="0 0 1440 320"
               preserveAspectRatio="xMidYMid slice"
-              style={{
-                position: 'absolute',
-                top: svgYOffset,
-                zIndex: 5,
-                backgroundColor: 'transparent',
-              }}
             >
               <Path
                 fill="#ffffff"
                 d="M 0,420 L 0,157.2 C 123.3,116.6 246.7,76.0 401,81.6 C 555.3,87.2 740.7,151.7 919,180.0 C 1097.3,208.3 1268.7,180.0 1540,61.2 L 1440,1320 L 0,1320 Z"
               />
-            </Svg>
-          </View>
+            </S.StyledSvg>
+          </S.TopTransparentSection>
 
-          <View style={styles.bottomOpaqueSection}>
-            <View
-              style={styles.content}
+          <S.BottomOpaqueSection>
+            <S.Content
               onLayout={(event) => {
                 contentHeight.value = event.nativeEvent.layout.height;
               }}
             >
                   {Array.isArray(children) ? children.map((child, index) => (
-                    <View key={index}>
+                    <S.EmptyContainer key={index}>
                         {child}
-                    </View>
+                    </S.EmptyContainer>
                     )) : (
-                    <View key={0}>
+                    <S.EmptyContainer key={0}>
                         {children}
-                    </View>
+                    </S.EmptyContainer>
                     )}
-            </View>
-          </View>
+            </S.Content>
+          </S.BottomOpaqueSection>
         </Animated.View>
       </GestureDetector>
     </GestureHandlerRootView>
@@ -169,20 +177,6 @@ const ParallaxDrawer = ({ headerImage, children, onClose }: ParallaxDrawerProps)
 export default ParallaxDrawer;
 
 const styles = StyleSheet.create({
-  container: {
-    backgroundColor: 'green',
-  },
-  backgroundContainer: {
-    position: 'absolute',
-    top: 0,
-    height: HALF_HEIGHT * 1.1,
-    width: '100%',
-    zIndex: 0,
-  },
-  backgroundImage: {
-    width: '100%',
-    height: '100%',
-  },
   drawer: {
     position: 'absolute',
     top: 0,
@@ -191,30 +185,5 @@ const styles = StyleSheet.create({
     zIndex: 2,
     // backgroundColor: 'blue'
     overflow: 'hidden',
-  },
-  topTransparentSection: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    height: moderateScale(100),
-    backgroundColor: 'rgba(255, 255, 255, 0)',
-    zIndex: 1,
-  },
-  bottomOpaqueSection: {
-    position: 'absolute',
-    top: moderateScale(75),
-    left: 0,
-    right: 0,
-    // backgroundColor: 'pink',
-    backgroundColor: '#fff',
-    zIndex: 3,
-  },
-  content: {
-    zIndex: 3,
-    paddingVertical: '5%',
-    alignItems: 'center',
-    // justifyContent: 'center',
-    // backgroundColor: 'green',
-  },
+  }
 });
